% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addConstantColumns.R
\name{addConstantColumns}
\alias{addConstantColumns}
\title{Add columns with constant values}
\usage{
addConstantColumns(x, ccols)
}
\arguments{
\item{x}{[\code{data.frame}]\cr
Data frame with at least one row and column.}

\item{ccols}{[\code{list}]\cr
Named list of single elements.}
}
\value{
[\code{data.frame}]
}
\description{
Add columns with constant values
}
\examples{
x = data.frame(a = 1:2, b = 3:4)
y = addConstantColumns(x, list(c = "a", d = 0.1))
}
