% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame2.R
\name{as.data.frame2}
\alias{as.data.frame2}
\title{Conversion to data frame}
\usage{
as.data.frame2(x, col.names = NULL, ...)
}
\arguments{
\item{x}{[\code{object}]\cr
Any R object that can be coerced into a data frame.}

\item{col.names}{[\code{character} | \code{NULL}]\cr
Optional character vector of column names.}

\item{...}{[\code{any}]\cr
Further arguments passed down to \code{\link[base]{as.data.frame}}.}
}
\value{
[\code{data.frame}]
}
\description{
Wrapper around \code{\link[base]{as.data.frame}}.
In contrast it sets \code{stringsAsFactors=FALSE} and allows to
pass a vectors of column names.
}
\examples{
x = list(a = 1:2, b = c("K", "L"))
y = as.data.frame2(x, col.names = c("C1", "C2"))
}
