% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_named_list.R
\name{as_named_list}
\alias{as_named_list}
\title{Convert to named list}
\usage{
as_named_list(x, names, sep = ".")
}
\arguments{
\item{x}{[\code{vector}]\cr
Vector to be converted.}

\item{names}{[\code{character}]\cr
Either a single character or vector of \code{length(x)} characters.}

\item{sep}{[\code{string}]\cr
Single character used to split name from consecutive numbers in terms
that \code{names} is a single character.}
}
\value{
A named list.
}
\description{
Convert to named list
}
\examples{
x = 1:3
as_named_list(x, names = "x", sep = "-")
as_named_list(x, names = "x")
as_named_list(x, names = c("e1", "e2", "e3"))
}
