% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asNamedList.R
\name{asNamedList}
\alias{asNamedList}
\title{Convert to named list}
\usage{
asNamedList(x, names, sep = ".")
}
\arguments{
\item{x}{[\code{vector}]\cr
Vector to be converted.}

\item{names}{[\code{character}]\cr
Either a single character or vector of \code{length(x)} characters.}

\item{sep}{[\code{string}]\cr
Single character used to split name from consecutive numbers in terms
that \code{names} is a single character.}
}
\value{
[\code{list}] Named list.
}
\description{
Convert to named list
}
\examples{
x = 1:3
asNamedList(x, names = "x", sep = "-")
asNamedList(x, names = "x")
asNamedList(x, names = c("e1", "e2", "e3"))
}
