% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbindlapply.R
\name{cbindlapply}
\alias{cbindlapply}
\alias{rbindlapply}
\title{Bind lapply results.}
\usage{
cbindlapply(X, FUN, ...)

rbindlapply(X, FUN, ...)
}
\arguments{
\item{X}{[any]\cr
Iterable objects, e.g. list or vector.}

\item{FUN}{[\code{function}]\cr
Function applied to each element of \code{X} (see \code{\link[base]{lapply}}).}

\item{...}{[any]\cr
Furhter parameters passed down to \code{\link[base]{lapply}}.}
}
\value{
See return value of \code{\link[base]{cbind}} and \code{\link[base]{rbind}}
respectively.\cr
}
\description{
Simple shortcut for \code{do.call(cbind, lapply(...))}.
}
