% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_add_category.R
\name{df_add_category}
\alias{df_add_category}
\title{Add category for column.}
\usage{
df_add_category(x, col, values, new.value, factor.handling = "keep")
}
\arguments{
\item{x}{[\code{data.frame}]\cr
Data frame with at least one row and column.}

\item{col}{[\code{integer} | \code{string}]\cr
Column number or colmumn name.}

\item{values}{[\code{character}]\cr
Which values of the column \code{col} shall be unioned into the new category.}

\item{new.value}{[\code{any}]\cr
New category value.}

\item{factor.handling}{[\code{string}]\cr
Determines how to handle the case that \code{col} is a factor column.
Option \dQuote{drop} converts the factor into a character column. Option
\dQuote{keep} does the same, but after adding the new category, the factor
property is reestablished. If the factor was ordered, the new category is placed
last.}
}
\value{
[\code{data.frame}]
}
\description{
Add category for column.
}
