% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_add_constant_columns.R
\name{df_add_constant_columns}
\alias{df_add_constant_columns}
\title{Add columns with constant values}
\usage{
df_add_constant_columns(x, ccols)
}
\arguments{
\item{x}{[\code{data.frame}]\cr
Data frame with at least one row and column.}

\item{ccols}{[\code{list}]\cr
Named list of single elements.}
}
\value{
Modified data frame.
}
\description{
Add columns with constant values
}
\examples{
x = data.frame(a = 1:2, b = 3:4)
y = df_add_constant_columns(x, list(c = "a", d = 0.1))
}
\seealso{
Other data frame helpers: 
\code{\link{df_add_category}()},
\code{\link{df_explode}()},
\code{\link{df_rows_to_list}()},
\code{\link{df_split_col}()}
}
\concept{data frame helpers}
