% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_rows_to_list.R
\name{df_rows_to_list}
\alias{df_rows_to_list}
\title{Convert data frame rows to list}
\usage{
df_rows_to_list(x, named = TRUE)
}
\arguments{
\item{x}{[\code{data.frame}]\cr
Data frame.}

\item{named}{[\code{logical(1)}]\cr
Should the sub-lists be named?
If \code{TRUE} the \code{colnames(x)} are used if available.
Otherwise names \dQuote{V1, ...,VX} are used where X equals
\code{ncol(x)}.
Default is \code{TRUE}.}
}
\value{
List of (named) lists.
}
\description{
Returns a list of lists with each sub-list being a
row of the data frame.
}
\examples{
x = data.frame(x = 1:3, y = letters[3:5])
df_rows_to_list(x)
df_rows_to_list(x, named = FALSE)
df_rows_to_list(unname(x), named = FALSE)
}
\seealso{
Other data frame helpers: 
\code{\link{df_add_category}()},
\code{\link{df_add_constant_columns}()},
\code{\link{df_explode}()},
\code{\link{df_split_col}()}
}
\concept{data frame helpers}
