% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_split_col.R
\name{df_split_col}
\alias{df_split_col}
\title{Split string/factor column}
\usage{
df_split_col(x, split.col, scheme, delim, append = TRUE)
}
\arguments{
\item{x}{[\code{data.frame}]\cr
Source data frame.}

\item{split.col}{[\code{string}]\cr
Column name of column used for splitting.}

\item{scheme}{[\code{string}]\cr
Single string encoding the names and types of the split parts.
E.g. \dQuote{group{c}/problem{c}_n{i}_m{i}/algorithm{c}_pc{r}}.
Passed down to \code{decode_scheme}.}

\item{delim}{[\code{character}]\cr
Set of characters used to split the entries in the column.}

\item{append}{[\code{logical(1)}]\cr
Should the splitted data be appended to \code{x} by \code{split.col}?
Default is \code{FALSE}.}
}
\value{
Modified data frame.
}
\description{
Given a data frame \code{x}, a column name and a naming
scheme, the function splits every element in \code{x[[split.col]]},
using the naming scheme. Next, either a data frame with column names
indicating the split parts for each unique value in \code{x[[split.col]]}
is returned, or, if \code{append=TRUE}, this data frame is left-joined
with \code{x} by \code{split.col}.
}
\seealso{
Other data frame helpers: 
\code{\link{df_add_category}()},
\code{\link{df_add_constant_columns}()},
\code{\link{df_explode}()},
\code{\link{df_rows_to_list}()}
}
\concept{data frame helpers}
