% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_dirname.R
\name{dirname2}
\alias{dirname2}
\title{k-fold dirname.}
\usage{
dirname2(path, depth = 1L)
}
\arguments{
\item{path}{[\code{character}]\cr
Paths.}

\item{depth}{[\code{integer(1)}]\cr
Depth of the k-fold}
}
\value{
[\code{character}]
}
\description{
Given paths and an integer depth, the function calculates
dirname(...(...dirname(paths))), depth times.
}
