% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_path_pop.R
\name{fs_path_pop}
\alias{fs_path_pop}
\title{Stack-like pop operation for path names.}
\usage{
fs_path_pop(path, depth = 1L)
}
\arguments{
\item{path}{[\code{character}]\cr
Paths.}

\item{depth}{[\code{integer(1)}]\cr
How many parts should be dropped?
Default to 1. In this case the function behaves like \code{basename(path)}.}
}
\value{
[\code{character}]
}
\description{
Stack-like pop operation for path names.
}
