% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave2.R
\name{ggsave2}
\alias{ggsave2}
\title{Export ggplot object to file}
\usage{
ggsave2(filename, plot = ggplot2::last_plot(), ...)
}
\arguments{
\item{filename}{[\code{character(1)}]\cr
Path to file.}

\item{plot}{[\code{ggplot}]\cr
ggplot object. Defaults to last plot.}

\item{...}{[any]\cr
Further parameters passed down to \code{\link[ggplot2]{ggsave}}.}
}
\value{
Nothing
}
\description{
Simple wrapper to \code{\link[ggplot2]{ggsave}} setting \code{limitsize=FALSE}
and \code{device=cairo_pdf} by default.
}
