% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Replace elements in a named vector or list}
\usage{
insert(x, y, which)
}
\arguments{
\item{x}{[\code{vector}]\cr
First vector.}

\item{y}{[\code{vector}]\cr
Second vector.}

\item{which}{[\code{character}]\cr
Entry names that should be replaced in \code{x}.
Defaults to \code{names(y)}.}
}
\value{
Either a vector or a list depending on the type of \code{x} and \code{y}.
}
\description{
Given two named vectors/lists \code{x} and \code{y} and a list of names
this function replaces all entries of \code{x} with the matching entries in \code{y}.
}
\examples{
x = list(a = 100, b = letters[1:3])
y = list(b = "hello", c = "world")
insert(x, y)
insert(y, x)
insert(x, y, which = "b")
}
