% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.permutation.R
\name{is.permutation}
\alias{is.permutation}
\title{Test for permutation.}
\usage{
is.permutation(x, s)
}
\arguments{
\item{x}{[\code{vector}]\cr
Input vector to be checked.}

\item{s}{[\code{vector}]\cr
Base set.}
}
\value{
[\code{logical(1)}]
}
\description{
Check whether \code{x} is a permutation of the multi-set \code{s}.
}
\examples{
is.permutation(shuffle(1:10), 1:10)
is.permutation(c(2, 1, 4), s = 1:4)
}
