% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.suset.R
\name{is.subset}
\alias{is.subset}
\alias{is.superset}
\title{Subset check.}
\usage{
is.subset(x, y, strict = FALSE)

is.superset(x, y, strict = FALSE)
}
\arguments{
\item{x}{[\code{vector}]\cr
First vector.}

\item{y}{[\code{vector}]\cr
Second vector.}

\item{strict}{[\code{logical}]}
}
\value{
[\code{logical(1)}]
}
\description{
Check if vector \code{x} is sub- or super-set of vector \code{y}.
}
\examples{
x = c(1, 2, 3, 6, 7, 8)
y = 1:10
is.subset(x, y)
is.subset(y, x)
is.superset(x, y)
is.superset(y, x)
}
