% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_permutation.R
\name{is_permutation}
\alias{is_permutation}
\title{Test for permutation}
\usage{
is_permutation(x, s)
}
\arguments{
\item{x}{[\code{vector}]\cr
Input vector to be checked.}

\item{s}{[\code{vector}]\cr
Base set.}
}
\value{
A scalar logical value.
}
\description{
Check whether \code{x} is a permutation of the multi-set \code{s}.
}
\examples{
is_permutation(shuffle(1:10), 1:10)
is_permutation(c(2, 1, 4), s = 1:4)
}
