% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_subset.R
\name{is_subset}
\alias{is_subset}
\alias{is_superset}
\title{Subset check}
\usage{
is_subset(x, y, strict = FALSE)

is_superset(x, y, strict = FALSE)
}
\arguments{
\item{x}{[\code{vector}]\cr
First vector.}

\item{y}{[\code{vector}]\cr
Second vector.}

\item{strict}{[\code{logical}]}
}
\value{
A single logical value.
}
\description{
Check if vector \code{x} is sub- or super-set of vector \code{y}.
}
\examples{
x = c(1, 2, 3, 6, 7, 8)
y = 1:10
is_subset(x, y)
is_subset(y, x)
is_superset(x, y)
is_superset(y, x)
}
