% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_list.R
\name{named_list}
\alias{named_list}
\title{Create a named list}
\usage{
named_list(ns, init = NULL)
}
\arguments{
\item{ns}{[\code{character}]\cr
Vector of list item names.}

\item{init}{[any]\cr
Initial value for each list item.
Defaults to \code{NULL}.}
}
\value{
A named list.
}
\description{
Generator for named list (length corresponds to the number of
names) where each item is optionally initialized with a pre-defined value.
}
\examples{
named_list(letters[1:10])
named_list(letters[1:10], init = NA)
named_list(letters[1:10], init = list())
named_list(letters[1:10], init = named_list(letters[1:2]))
}
