% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedList.R
\name{namedList}
\alias{namedList}
\title{Create a named list.}
\usage{
namedList(ns, init = NULL)
}
\arguments{
\item{ns}{[\code{character}]\cr
Vector of list item names.}

\item{init}{[any]\cr
Initial value for each list item.
Defaults to \code{NULL}.}
}
\value{
[\code{list}]
}
\description{
Generator for named list (length corresponds to the number of
names) where each item is optionally initialized with a pre-defined value.
}
\examples{
namedList(letters[1:10])
namedList(letters[1:10], init = NA)
namedList(letters[1:10], init = list())
namedList(letters[1:10], init = namedList(letters[1:2]))
}
