% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nunique.R
\name{nunique}
\alias{nunique}
\title{Get number of unique elements.}
\usage{
nunique(x)
}
\arguments{
\item{x}{[\code{atomic} | \code{factor}]\cr
A vector.}
}
\value{
[\code{integer(1)}]
}
\description{
Given a vector of atomic type (\code{logical}, \code{integer},
\code{numeric}, \code{complex}, \code{character}) or a factor the function
returns the number of different elements.
}
\note{
NAs are counted as a seperate value.
}
\examples{
x = c(1, 2, 2, 3, 2, NA, NA, 5)
nunique(x) # 5
}
