% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{re_compile}
\alias{re_compile}
\title{re_compile}
\usage{
re_compile(
  pattern,
  IGNORECASE,
  I,
  MULTILINE,
  M,
  DOTALL,
  S,
  VERBOSE,
  X,
  NOFLAG,
  ...
)
}
\arguments{
\item{pattern}{string Regular expression pattern.}

\item{IGNORECASE}{boolean Whether or not to ignore case during matching.}

\item{I}{boolean Alternative to IGNORECASE.}

\item{MULTILINE}{boolean whether or not the matches should span over multiple
lines.}

\item{M}{boolean ALternative for MULTILINE.}

\item{DOTALL}{boolean Whether \code{.} should match line breaks.}

\item{S}{boolean Alternative for DOTALL.}

\item{VERBOSE}{boolean Whether pattern definitions span multiple lines.}

\item{X}{boolean Alternative for VERBOSE.}

\item{NOFLAG}{boolean If this flag is passed, all others are ignored.}
}
\description{
Roughly equivalent to python's re.compile. Though, it doesn't
provide the same performance benefits as python's re.compile.
}
