% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{re_sub}
\alias{re_sub}
\title{re_sub}
\usage{
re_sub(pattern, replacement, string, ...)
}
\arguments{
\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described
in \link[stringi:about_search_regex]{stringi::about_search_regex}. Control options with
\code{\link[stringr:regex]{regex()}}.

For \code{str_replace_all()} this can also be a named vector
(\code{c(pattern1 = replacement1)}), in order to perform multiple replacements
in each element of \code{string}.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:coll]{coll()}} which
respects character matching rules for the specified locale.}

\item{replacement}{The replacement value, usually a single string,
but it can be the a vector the same length as \code{string} or \code{pattern}.
References of the form \verb{\\1}, \verb{\\2}, etc will be replaced with
the contents of the respective matched group (created by \verb{()}).

Alternatively, supply a function, which will be called once for each
match (from right to left) and its return value will be used to replace
the match.}

\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{...}{
  Arguments passed on to \code{\link[=re_compile]{re_compile}}
  \describe{
    \item{\code{IGNORECASE}}{boolean Whether or not to ignore case during matching.}
    \item{\code{I}}{boolean Alternative to IGNORECASE.}
    \item{\code{MULTILINE}}{boolean whether or not the matches should span over multiple
lines.}
    \item{\code{M}}{boolean ALternative for MULTILINE.}
    \item{\code{DOTALL}}{boolean Whether \code{.} should match line breaks.}
    \item{\code{S}}{boolean Alternative for DOTALL.}
    \item{\code{VERBOSE}}{boolean Whether pattern definitions span multiple lines.}
    \item{\code{X}}{boolean Alternative for VERBOSE.}
    \item{\code{NOFLAG}}{boolean If this flag is passed, all others are ignored.}
  }}
}
\description{
Replace all occurrences of \code{pattern} in \code{string} with
\code{replacement}.
}
