% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample2.R
\name{sample2}
\alias{sample2}
\title{Modified version of \code{\link[base]{sample}}.}
\usage{
sample2(x, size, replace = FALSE, ...)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose,
    or a positive integer.  See \sQuote{Details.}}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{...}{[any]\cr
Further parameters passed down to \code{\link[base]{sample}}, e.g. \code{prob}.}
}
\value{
[\code{vector}]
}
\description{
Function \code{\link[base]{sample}}, given the first parameter
\code{x} is a single integer value, samples from the set \eqn{1,\ldots,x}.
This behavior might be undesirable when \code{\link[base]{sample}} is embedded
into routines and \code{x} is a variable.
This function remedies this issue. If \code{x} is a number it returns \code{x}
\code{size}-times.
}
