% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{Random shuffle of a vector.}
\usage{
shuffle(x)
}
\arguments{
\item{x}{[\code{vector}]\cr
Input vector.}
}
\value{
[\code{vector}] Shuffled/scrambled \code{x}.
}
\description{
Basically an alias for \code{\link[base]{sample}}
Note that if the \code{x} is a scalar value the function returns
simply the value itself, i.e. acts like the identity function.
}
\examples{
shuffle(1:10)
}
