% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_explode.R
\name{str_explode}
\alias{str_explode}
\alias{explode}
\title{Split strings into parts of certain types}
\usage{
str_explode(x, split, names = NULL, types = NULL)

explode(x, split, names = NULL, types = NULL)
}
\arguments{
\item{x}{[\code{character}]\cr
Character vector.}

\item{split}{[\code{character(1)}]\cr
See parameter \code{split} of \code{\link[base]{strsplit}}.}

\item{names}{[\code{character}]\cr
Optional column names.
If \code{NULL}, the column names are explode1 to explodeN where N is the number of
split parts.}

\item{types}{[\code{character(1)}]\cr
Character string where the i-th character specifies the atomic data type of
the i-th split part. Possible characters are \dQuote{'c'} (character),
\dQuote{'f'} (factor), dQuote{'i'} (integer), \dQuote{'l'} (logical) or
\dQuote{n}, \dQuote{d} or \dQuote{r} for numeric/double/real.
Default is \code{NULL}. In this case all split parts are characters.}
}
\value{
A data frame.
}
\description{
Extract information from strings. This is done by splitting the
string by one or multiple delimiters. The split results are wrapped in a
\code{data.frame}. Optionally, one can pass column names and/or types for the
parts. E.g. \code{types="cir"} converts the first part into character, the
second into integer and the third into real/double.
}
\seealso{
Other string helpers: 
\code{\link{str_parse}()},
\code{\link{str_to_shortcut}()}
}
\concept{string helpers}
