% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_shortcut.R
\name{str_to_shortcut}
\alias{str_to_shortcut}
\title{Make string shortcuts}
\usage{
str_to_shortcut(x, split, nchars = 1, sep = "-", ...)
}
\arguments{
\item{x}{[\code{character}]
Vector of strings.}

\item{split}{[\code{character(1)}]\cr
Character string used to split each component of \code{x}.}

\item{nchars}{[\code{integer(1)}]\cr
Length of extracted prefix for each part.
Defaults to 1.}

\item{sep}{[\code{string}]\cr
Possibly empty glue string.
Default is \dQuote{-}.}

\item{...}{[\code{any}]\cr
Further arguments passed down to \code{\link[base]{strsplit}}.
The most important here is \code{split}.}
}
\value{
Shortcut version of \code{x}.
}
\description{
Given a character vector, each element is split, the first
\code{nchars} characters are extracted from each part and all parts are
concatenated with \code{sep}.
This is useful to obtain shortcuts of lengthy strings.
}
\examples{
x = c("hello-world", "some-length-string")
str_to_shortcut(x, split = "-")
str_to_shortcut(x, nchars = 2L, sep = "", split = "-")
}
\seealso{
Other string helpers: 
\code{\link{str_explode}()},
\code{\link{str_parse}()}
}
\concept{string helpers}
