% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symdiff.R
\name{symdiff}
\alias{symdiff}
\title{Set operation symmetric difference.}
\usage{
symdiff(x, y)
}
\arguments{
\item{x}{[\code{vector}]\cr
Vector containing a sequence of items.}

\item{y}{[\code{vector}]\cr
Another vector containing a sequence of items.}
}
\value{
[\code{vector}]
}
\description{
Performs symmetric(!) \textbf{set} difference in contrast
to \code{\link[base]{setdiff}}. I.e. given two vectors (of the same mode) the result
contains all elements that are either in the first or the second vector
but not in both. Duplicated values in the arguments will be discarded.
}
\examples{
symdiff(1:10, 6:15)
symdiff(1:10, 1:20)
symdiff(c("a", "b", "c"), c("d", "c", "c", "d", "e"))
}
