% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_min.R
\name{which_min}
\alias{which_min}
\alias{which_max}
\title{Where is is min or max value in a nmeric vector?}
\usage{
which_min(x, return.method = "first")

which_max(x, return.method = "first")
}
\arguments{
\item{x}{[\code{numeric}]\cr
Numeric vector.}

\item{return.method}{[\code{string}]\cr
What to return in the presence of ties?
\itemize{
\item{\dQuote{first}: first occurence}
\item{\dQuote{last}: last occurence}
\item{\dQuote{random}: sampled uniformly at random from all positions}
\item{\dQuote{all}: returns a vector of all positions}
}
Default is \dQuote{first}.}
}
\value{
Vector of integer position(s).
}
\description{
Where is is min or max value in a nmeric vector?
}
\examples{
x = c(10, 24, 2, 2, 15, 2, 28)
lapply(c("first", "last", "random", "all"), function(m) {
 which_min(x, return.method = m)
})
}
