% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip}
\alias{zip}
\title{Aggregate vectors/lists into lists of lists.}
\usage{
zip(...)
}
\arguments{
\item{...}{[Lists/vectors]\cr
A variable number of lists or vectors of the same size.}
}
\value{
[\code{list}]
}
\description{
This function emulates Python's zip function. It expects a
variable number of arguments. If a single vector/list x is passed the
return value is a list of lists where each sub-list contains the components
\dQuote{key} and \dQuote{value}. If x is named the key is the name, otherwise
it is the element number. The value field unsuprisingly contains the
corresponding value.
}
\examples{
x = list("a" = 1, "b" = 2, "c" = 3)
for (el in zip(x)) {
  catf("Key: '\%s', Value: \%i\\n", el$key, el$value)
}

z = zip(x, x)
}
