% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse string}
\usage{
collapse(x, sep = ",")
}
\arguments{
\item{x}{[\code{vector}]\cr
Input vector.}

\item{sep}{[\code{character(1)}]\cr
Character used to seperate the entries of \code{x} in the produced string.}
}
\value{
[\code{character(1)}]
}
\description{
Collapse string
}
