% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullmatch.R
\name{re_fullmatch}
\alias{re_fullmatch}
\title{Match a pattern against the entire string}
\usage{
re_fullmatch(pattern, string, ...)
}
\arguments{
\item{pattern}{A regular expression pattern or a compiled pattern object.}

\item{string}{A character vector where each element is a string to be matched
against the pattern.}

\item{...}{Additional arguments to be passed to \code{re_compile} if the pattern
needs to be compiled.}
}
\value{
A list where each element is a character vector containing the full
match for the corresponding element of \code{string}, or \code{character(0)} if there
is no match.
#' @inheritDotParams re_compile
}
\description{
\code{re_fullmatch} checks whether each element of a character vector fully
matches a specified pattern (regular expression). If the provided pattern is
not already a compiled pattern object, it compiles it using \code{re_compile}. The
function ensures that the entire string matches the pattern from start to
end.
}
\examples{
pattern <- re_compile("\\\\d{3}-\\\\d{2}-\\\\d{4}")
re_fullmatch(pattern, "123-45-6789") # Full match
re_fullmatch("123-45-6789", "123-45-6789 and more") # No full match
}
\seealso{
\href{https://docs.python.org/3/library/re.html#re.fullmatch}{Python re.fullmatch() documentation}
}
