% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{re_split}
\alias{re_split}
\title{re_split}
\usage{
re_split(pattern, string, ..., n = Inf)
}
\arguments{
\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described in
\code{vignette("regular-expressions")}. Use \code{\link[stringr:regex]{regex()}} for finer control of the
matching behaviour.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:coll]{coll()}} which
respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with
\code{\link[stringr:boundary]{boundary()}}. An empty pattern, "", is equivalent to
\code{boundary("character")}.}

\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{...}{
  Arguments passed on to \code{\link[=re_compile]{re_compile}}
  \describe{
    \item{\code{IGNORECASE}}{boolean Whether or not to ignore case during matching.}
    \item{\code{I}}{boolean Alternative to IGNORECASE.}
    \item{\code{MULTILINE}}{boolean whether or not the matches should span over multiple
lines.}
    \item{\code{M}}{boolean ALternative for MULTILINE.}
    \item{\code{DOTALL}}{boolean Whether \code{.} should match line breaks.}
    \item{\code{S}}{boolean Alternative for DOTALL.}
    \item{\code{VERBOSE}}{boolean Whether pattern definitions span multiple lines.}
    \item{\code{X}}{boolean Alternative for VERBOSE.}
    \item{\code{NOFLAG}}{boolean If this flag is passed, all others are ignored.}
  }}

\item{n}{Maximum number of pieces to return. Default (Inf) uses all
possible split positions.

For \code{str_split()}, this determines the maximum length of each element
of the output. For \code{str_split_fixed()}, this determines the number of
columns in the output; if an input is too short, the result will be padded
with \code{""}.}
}
\description{
Split \code{string} on \code{pattern}. Split will occur \code{n} times.
}
